backend/requirements.txt
fastapi>=0.104.0
sqlmodel>=0.0.14
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.24.0

